refactor(persona): split persona.py into focused modules (RFC 0005, PR 8a)
Extract four new modules from the 2,091-line persona.py:

- agents/persona_types.py (229 lines): EventType, AgentEvent, ActionType,
  AgentAction, SubAgentRequest, SubAgentResult, OrchestratorClient protocol,
  Mood enum, PersonaState dataclass with energy constants.

- agents/persona_behavior.py (131 lines): DIMENSION_DESCRIPTIONS mapping,
  _DIMENSION_DEFAULTS, dimension key-set assertion, render_behavior().

- agents/dispatch.py (422 lines): ActionExecutor, EventDispatcher classes
  with _MAX_MENTIONS_PER_ACTION, _DEFAULT_DISPATCH_TIMEOUT constants.
  Uses TYPE_CHECKING for _LLMPersonaAgent and TickScheduler to avoid
  circular imports.

- agents/tick.py (219 lines): TickScheduler class.
  Uses TYPE_CHECKING for _LLMPersonaAgent and ActionExecutor.

persona.py retains PersonaAgent ABC, _LLMPersonaAgent, create_persona_agent()
factory, _truncate_with_ellipsis(), _coerce_event_timeout(), and constants
used exclusively by _LLMPersonaAgent (~1,190 lines -- larger than the 650
target because _LLMPersonaAgent grew substantially during PR 7b review rounds).

All symbols remain importable from agents.persona via re-exports.
agents/__init__.py updated to import from canonical locations.
agents/server.py updated to import from new modules.
No test changes required -- all existing import paths preserved.

814 passed, 2 skipped, ruff clean, no circular imports.
