+++ FILEMAP.md
@@ -1,14 +1,14 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-14  
-> **183** tracked files
+> **187** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
 | `internal/` | 40 |
-| `agents/` | 28 |
+| `agents/` | 32 |
 | `tests/` | 28 |
 | `docs/` | 21 |
 | `(root)` | 19 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 65 |
+| `.py` | 69 |
 | `.md` | 35 |
 | `.go` | 35 |
 | `.yaml` | 22 |
@@ -79,6 +79,7 @@
 ├── agents/
 │   ├── __init__.py
 │   ├── base.py
+│   ├── dispatch.py
 │   ├── generated/
 │   │   ├── __init__.py
 │   │   ├── agent_message_pb2.py
@@ -93,6 +94,8 @@
 │   │   ├── relationship.py
 │   │   └── working.py
 │   ├── persona.py
+│   ├── persona_behavior.py
+│   ├── persona_types.py
 │   ├── pyproject.toml
 │   ├── server.py
 │   ├── sub_agents/
@@ -102,6 +105,7 @@
 │   ├── task_agent.py
 │   ├── tests/
 │   │   └── __init__.py
+│   ├── tick.py
 │   ├── tools/
 │   │   ├── __init__.py
 │   │   ├── builtin.py
